software-engineering/backend/pyproject.toml
[tool.poetry]
name = "inspektor"
version = "0.1.0"
description = "Small project for Software Engineer job interview"
authors = ["Adriano Pagano <adriano.pagano@vu.engineering>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "inspektor"}]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
django-storages = {extras = ["azure"], version = "^1.14.2"}
django-filter = "^23.5"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
tensorflow = "^2.15"
whitenoise = "^6.6.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
